+++ .env.sample
@@ -1,5 +1,6 @@
-AMAZON_API_KEY=
-AMAZON_API_SECRET=
+AMAZON_CREATOR_API_ID=
+AMAZON_CREATOR_API_SECRET=
+AMAZON_CREATOR_API_VERSION=
 AMAZON_TAG=
 AMAZON_COUNTRY=
 BITLY_API_TOKEN=
