+++ pyproject.toml
@@ -4,11 +4,11 @@ version = "0.1.0"
 description = ""
 authors = [{name = "hito_horobe"}]
 readme = "README.md"
-requires-python = ">=3.11"
+requires-python = ">=3.13"
 dependencies = [
     "beautifulsoup4>=4.12.3",
     "requests>=2.32.3",
-    "python-amazon-paapi>=5.0.1",
+    "python-amazon-paapi>=6.2.0",
     "requests-cache>=1.2.1",
     "pydantic>=2.10.6",
     "python-dotenv>=1.0.1",
